adjust URL quoting, only quote when there is `%` in URL
This is prompted by some Wikipedia URL which use parentheses, which requires quoting.

If the URL is HTTP, then Wikipedia would redirect to HTTPS, and the URL will be quoted again, resulting 400.

This is just a simple workaround, lnkckr should not do any quoting for the user, that is the input URLs are already valid URL, but since browsers seem to take care of such already, this provides some convenience.